Remove obsolete redirect from ProjectWizardController
redirect_if_blank_client was used to redirect to
project_wizard_path(:add_project) if current_user.github_client was blank?
As we inherit from the DashboardController this code becomes obsolete,
as there is logic in the DashboardController#check_for_active_providers
that handles these redirects when github_client is blank.
The only difference is that we redirect to root_path instead of
redirecting to project_wizard_path(:add_project)