Run chunks with heaviest test first
We start filling chunks starting with the heaviest jobs. This means, the
smaller the chunk index the heavier the test jobs. This way the heaviest
jobs start early so they have the maximum available time to finish.
E.g. A suite has a job that takes 3 minutes to complete and the rest
of the suite needs 3 minutes. If the heavy test does not start
immediately, the total time will be 3 minutes plus the delay to start.