Skip TestRun setup when test_run is cancelled
Adding a new testrun on a tracked branch, cancels all pending. This
might happen before the job is run or in the middle of the job. To avoid
resetting the status back to non-cancelled we skip save! and the rest of
the method when we find a cancelled status on the database.

We could create a transaction but this method might take too long to
complete.